Code adapted (Graph Airline distance) so that Heuristic of airline Distance is admissible (Edgeweight is greater or equal than airlineDistance).

+++ src/main/java/customVertex.java
@@ -3,7 +3,7 @@ public class customVertex{
 
     public int pathCost;
 
-    public int heuristicCosts;
+    public double heuristicCosts;
 
     public customVertex parentVertex;
